pyproject.toml
[project]
name = "svg_tool"
version = "0.1.0"
description = "Python scripts for playing with svg files"
readme = "README.md"
requires-python=">=3.8"
license = {file = "LICENSE.txt"}
keywords = []
authors = [
    {name="snake-biscuits", email="bikkie.dev@gmx.com"}
]
maintainers = [
    {name="snake-biscuits", email="bikkie.dev@gmx.com"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3 :: Only",
    "Topic :: Software Development :: Libraries :: Python Modules"
]
dependencies = [
    "lxml",
    "Pillow"
]


[project.optional-dependencies]
# NOTE: install with pip extras syntax:
# -- `$ pip install svg_tool[test]`
test = ["pytest", "pytest_cov"]


[project.urls]
"Repository" = "https://github.com/snake-biscuits/svg_tool"
"Changelog" = "https://github.com/snake-biscuits/svg_tool/blob/master/CHANGELOG.md"
"Bug Reports" = "https://github.com/snake-biscuits/svg_tool/issues"
"Funding" = "https://ko-fi.com/bikkie"


[build-system]
requires = ["hatchling >= 1.13"]
build-backend = "hatchling.build"


# hatchling config
[tool.hatch.build.targets.sdist]
include = [
    "/svg_tool/*",
    "/tests",
    "/CHANGELOG.md",
]


[tool.hatch.build.targets.wheel]
only-packages = true
